HASHMAP: Rearrange lib.rs and create hashmap module
